requirements.txt
tensorflow>=2.10
numpy
opencv-python
scikit-learn
matplotlib
seaborn
tifffile
Pillow
pyyaml
pytest
